Split p_m_serde_macros into modules
